+++ packages/terra-date-picker/tests/jest/DatePicker.test.jsx
@@ -1,10 +1,9 @@
 import React from 'react';
 /* eslint-disable import/no-extraneous-dependencies */
-import { IntlProvider } from 'react-intl';
+import { shallowWithIntl } from 'terra-enzyme-intl';
 import moment from 'moment';
 import DatePicker from '../../lib/DatePicker';
 import DateUtil from '../../lib/DateUtil';
-import messages from '../../translations/en-US.json';
 
 // Mock the following functions so that they always return a consistent date.
 // Otherwise, a date with a different offset would be created based on the the timezone where the tests are executed.
@@ -13,102 +12,68 @@ DateUtil.filterInvalidDates = jest.fn();
 DateUtil.createSafeDate.mockImplementation(() => moment.utc('2017-01-01'));
 DateUtil.filterInvalidDates.mockImplementation(() => [moment.utc('2017-01-01')]);
 
-const locale = 'en-US';
 const isWeekday = (date) => {
   const day = date.day();
   return day !== 0 && day !== 6;
 };
 
 it('should render a default date input and date picker', () => {
-  const datePicker = shallow((
-    <IntlProvider locale={locale} messages={messages}>
-      <DatePicker name="date-input" utcOffset={0} />
-    </IntlProvider>));
+  const datePicker = shallowWithIntl(<DatePicker name="date-input" utcOffset={0} />);
   expect(datePicker).toMatchSnapshot();
 });
 
 it('should render a default date input with custom input attributes', () => {
-  const datePicker = shallow((
-    <IntlProvider locale={locale} messages={messages}>
-      <DatePicker name="date-input" utcOffset={0} inputAttributes={{ id: 'terra-date-input' }} />
-    </IntlProvider>));
+  const datePicker = shallowWithIntl(<DatePicker name="date-input" utcOffset={0} inputAttributes={{ id: 'terra-date-input' }} />);
   expect(datePicker).toMatchSnapshot();
 });
 
 it('should render a date picker with filtered dates', () => {
-  const datePicker = shallow((
-    <IntlProvider locale={locale} messages={messages}>
-      <DatePicker filterDate={isWeekday} name="date-input" utcOffset={0} />
-    </IntlProvider>));
+  const datePicker = shallowWithIntl(<DatePicker filterDate={isWeekday} name="date-input" utcOffset={0} />);
   expect(datePicker).toMatchSnapshot();
 });
 
 it('should render a date picker with disabled dates', () => {
-  const datePicker = shallow((
-    <IntlProvider locale={locale} messages={messages}>
-      <DatePicker name="date-input" excludeDates={['2017-04-01']} utcOffset={0} />
-    </IntlProvider>));
+  const datePicker = shallowWithIntl(<DatePicker name="date-input" excludeDates={['2017-04-01']} utcOffset={0} />);
   expect(datePicker).toMatchSnapshot();
 });
 
 it('should render a date picker with included dates', () => {
-  const datePicker = shallow((
-    <IntlProvider locale={locale} messages={messages}>
-      <DatePicker name="date-input" includeDates={['2017-04-01']} utcOffset={0} />
-    </IntlProvider>));
+  const datePicker = shallowWithIntl(<DatePicker name="date-input" includeDates={['2017-04-01']} utcOffset={0} />);
   expect(datePicker).toMatchSnapshot();
 });
 
 it('should render a date picker with min and max dates', () => {
-  const datePicker = shallow((
-    <IntlProvider locale={locale} messages={messages}>
-      <DatePicker name="date-input" minDate="2017-04-01" maxDate="2017-04-10" utcOffset={0} />
-    </IntlProvider>));
+  const datePicker = shallowWithIntl(<DatePicker name="date-input" minDate="2017-04-01" maxDate="2017-04-10" utcOffset={0} />);
   expect(datePicker).toMatchSnapshot();
 });
 
 it('should render a date picker with releaseFocus', () => {
   const releaseFocus = () => {};
-  const datePicker = shallow((
-    <IntlProvider locale={locale} messages={messages}>
-      <DatePicker name="date-input" releaseFocus={releaseFocus} utcOffset={0} />
-    </IntlProvider>));
+  const datePicker = shallowWithIntl(<DatePicker name="date-input" releaseFocus={releaseFocus} utcOffset={0} />);
   expect(datePicker).toMatchSnapshot();
 });
 
 it('should render a date picker with requestFocus', () => {
   const requestFocus = () => {};
-  const datePicker = shallow((
-    <IntlProvider locale={locale} messages={messages}>
-      <DatePicker name="date-input" requestFocus={requestFocus} utcOffset={0} />
-    </IntlProvider>));
+  const datePicker = shallowWithIntl(<DatePicker name="date-input" requestFocus={requestFocus} utcOffset={0} />);
   expect(datePicker).toMatchSnapshot();
 });
 
 it('should render a date picker with onChange', () => {
   const onChange = () => {};
-  const datePicker = shallow((
-    <IntlProvider locale={locale} messages={messages}>
-      <DatePicker name="date-input" onChange={onChange} utcOffset={0} />
-    </IntlProvider>));
+  const datePicker = shallowWithIntl(<DatePicker name="date-input" onChange={onChange} utcOffset={0} />);
   expect(datePicker).toMatchSnapshot();
 });
 
 it('should render a date picker with onChangeRaw', () => {
   const onChangeRaw = () => {};
-  const datePicker = shallow((
-    <IntlProvider locale={locale} messages={messages}>
-      <DatePicker name="date-input" onChangeRaw={onChangeRaw} utcOffset={0} />
-    </IntlProvider>));
+  const datePicker = shallowWithIntl(<DatePicker name="date-input" onChangeRaw={onChangeRaw} utcOffset={0} />);
   expect(datePicker).toMatchSnapshot();
 });
 
 it('should render a date picker with onSelect', () => {
   const onSelect = () => {};
-  const datePicker = shallow((
-    <IntlProvider locale={locale} messages={messages}>
-      <DatePicker name="date-input" onSelect={onSelect} utcOffset={0} />
-    </IntlProvider>));
+  const datePicker = shallowWithIntl(<DatePicker name="date-input" onSelect={onSelect} utcOffset={0} />);
   expect(datePicker).toMatchSnapshot();
 });
 
@@ -124,27 +89,18 @@ it('throws error on missing locale prop in Base', () => {
 
 it('should render a date picker with onCalendarButtonClick', () => {
   const handleOnCalendarButtonClick = () => {};
-  const datePicker = shallow((
-    <IntlProvider locale={locale} messages={messages}>
-      <DatePicker name="date-input" onCalendarButtonClick={handleOnCalendarButtonClick} utcOffset={0} />
-    </IntlProvider>));
+  const datePicker = shallowWithIntl(<DatePicker name="date-input" onCalendarButtonClick={handleOnCalendarButtonClick} utcOffset={0} />);
   expect(datePicker).toMatchSnapshot();
 });
 
 it('should render a date picker with onInputFocus', () => {
   const handleOnInputFocus = () => {};
-  const datePicker = shallow((
-    <IntlProvider locale={locale} messages={messages}>
-      <DatePicker name="date-input" onInputFocus={handleOnInputFocus} utcOffset={0} />
-    </IntlProvider>));
+  const datePicker = shallowWithIntl(<DatePicker name="date-input" onInputFocus={handleOnInputFocus} utcOffset={0} />);
   expect(datePicker).toMatchSnapshot();
 });
 
 it('should render a disabled date picker', () => {
-  const datePicker = shallow((
-    <IntlProvider locale={locale} messages={messages}>
-      <DatePicker name="date-input" disabled utcOffset={0} />
-    </IntlProvider>));
+  const datePicker = shallowWithIntl(<DatePicker name="date-input" disabled utcOffset={0} />);
   expect(datePicker).toMatchSnapshot();
 });