+++ packages/terra-application-links/CHANGELOG.md
@@ -4,6 +4,12 @@ Changelog
 Unreleased
 ----------
 
+5.6.0 - (March 20, 2019)
+------------------
+### Changed
+* Update Jest tests to use terra-enzyme-intl
+* Changed demo mock background colors to be themable
+
 5.5.0 - (March 12, 2019)
 ------------------
 ### Changed