+++ packages/terra-date-picker/CHANGELOG.md
@@ -7,6 +7,11 @@ Unreleased
 * Added 'required' prop to match terra-form components' required prop
 git
 
+3.9.0 - (March 20, 2019)
+------------------
+### Changed
+* Update Jest tests to use terra-enzyme-intl
+
 3.8.0 - (March 12, 2019)
 ------------------
 ### Changed