+++ packages/terra-notification-dialog/CHANGELOG.md
@@ -4,6 +4,11 @@ Changelog
 Unreleased
 ----------
 
+2.6.0 - (March 20, 2019)
+------------------
+### Changed
+* Update Jest tests to use terra-enzyme-intl
+
 2.5.0 - (March 12, 2019)
 ------------------
 ### Added