+++ packages/terra-popup/CHANGELOG.md
@@ -3,8 +3,12 @@ Changelog
 
 Unreleased
 ----------
+
+5.6.0 - (March 20, 2019)
+------------------
 ### Changed
 * Updated heading levels in doc site README.md to correct Section 508 errors
+* Update Jest tests to use terra-enzyme-intl
 
 5.5.0 - (March 12, 2019)
 ------------------